fix NumericExtensions Max(int) and Min(int) to return int instead of float
git-svn-id: svn://srv-dc/svn2/Citrus@226179 20bf7d6a-bd21-9d4d-a3c4-a2c14bb8c68a

+++ Lime/Source/NumericExtensions.cs
@@ -22,12 +22,12 @@ namespace Lime
 			return a < b ? a : b;
 		}
 
-		public static float Max(this int a, int b)
+		public static int Max(this int a, int b)
 		{
 			return a > b ? a : b;
 		}
 
-		public static float Min(this int a, int b)
+		public static int Min(this int a, int b)
 		{
 			return a < b ? a : b;
 		}
